+++ vsd/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.2] - 2025-08-12
+
 ### Changed
 
 - `save`
@@ -197,7 +199,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2022-06-22
 
-[Unreleased]: https://github.com/clitic/vsd/compare/vsd-0.4.`...HEAD
+[Unreleased]: https://github.com/clitic/vsd/compare/vsd-0.4.2...HEAD
+[0.4.2]: https://github.com/clitic/vsd/compare/vsd-0.4.1...vsd-0.4.2
 [0.4.1]: https://github.com/clitic/vsd/compare/vsd-0.4.0...vsd-0.4.1
 [0.4.0]: https://github.com/clitic/vsd/compare/0.3.3...vsd-0.4.0
 [0.3.3]: https://github.com/clitic/vsd/compare/0.3.2...0.3.3